Patch tornado for .NET websocket compatibility.
This adds a generic patching mechanism to `kaggle/python` and a single patch
to apply the equivalent of https://github.com/tornadoweb/tornado/pull/2098 so
we don't have to wait for the upstream change to trickle down.

Resulting image has been verified to prevent kaggle websockets to be broken off
every 100s.