Adding the pykoko package
Koko is an information extraction tool (developed in Python 3) that allows users to query a text corpus and extract those entities that is of interest to them. A wide range of lexical, syntactic and semantic patterns can be specified using Koko’s query language which can be combined to efficiently extract and score relevant entities.